pyproject.toml
[project]
name = "walled-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.6.1",
    "openai-agents>=0.4.2",
    "python-dotenv>=1.1.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.1",
]
